+++ src/Client/W3CClient.php
@@ -62,7 +62,9 @@ class W3CClient implements ClientInterface
      *
      * ```
      * $loop = \React\EventLoop\Factory::create();
+     *
      * $browser = new \React\Http\Browser($loop);
+     * $browser = $browser->withRejectErrorResponse(false);
      *
      * $hubClient = new \Itnelo\React\WebDriver\Client\W3CClient(
      *     $browser,
@@ -114,7 +116,7 @@ class W3CClient implements ClientInterface
      */
     public function getSessionIdentifiers(): PromiseInterface
     {
-        // todo
+        // todo: implementation
 
         return reject(new RuntimeException('Not implemented.'));
     }
@@ -149,8 +151,8 @@ class W3CClient implements ClientInterface
                     preg_match('/sessionid[":\s]+([a-z\d]{32})/Ui', $responseBody, $matches);
 
                     if (!isset($matches[1])) {
-                        // todo: locate an error message or set it as "undefined error".
-                        throw new RuntimeException('Unable to locate session identifier in the response.');
+                        // todo: locate an error message or set it as "undefined error"
+                        throw new RuntimeException('Unable to locate a session identifier in the response.');
                     }
 
                     $sessionIdentifier = $matches[1];
@@ -182,7 +184,7 @@ class W3CClient implements ClientInterface
      */
     public function removeSession(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement removeSession() method.
+        // todo: implementation
 
         return reject(new RuntimeException('Not implemented.'));
     }
@@ -192,9 +194,47 @@ class W3CClient implements ClientInterface
      */
     public function getTabIdentifiers(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement getTabIdentifiers() method.
+        $tabLookupDeferred = new Deferred();
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/window/handles',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
+
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $responsePromise = $this->httpClient->get($requestUri, $requestHeaders);
+
+        $responsePromise->then(
+            function (ResponseInterface $response) use ($tabLookupDeferred) {
+                try {
+                    $tabIdentifiers = $this->deserializeResponse($response);
+
+                    $tabLookupDeferred->resolve($tabIdentifiers);
+                } catch (Throwable $exception) {
+                    $reason = new RuntimeException(
+                        'Unable to get tab identifiers (response deserialization).',
+                        0,
+                        $exception
+                    );
+
+                    $tabLookupDeferred->reject($reason);
+                }
+            },
+            function (Throwable $rejectionReason) use ($tabLookupDeferred) {
+                $reason = new RuntimeException('Unable to get tab identifiers (request).', 0, $rejectionReason);
+
+                $tabLookupDeferred->reject($reason);
+            }
+        );
+
+        $identifierListPromise = $tabLookupDeferred->promise();
+
+        return $identifierListPromise;
     }
 
     /**
@@ -202,9 +242,36 @@ class W3CClient implements ClientInterface
      */
     public function getActiveTabIdentifier(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement getActiveTabIdentifier() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/window',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $responsePromise = $this->httpClient->get($requestUri, $requestHeaders);
+
+        $tabIdentifierPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $tabIdentifier = $this->deserializeResponse($response);
+
+                    return $tabIdentifier;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to get an identifier for the active tab.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $tabIdentifierPromise;
     }
 
     /**
@@ -212,9 +279,38 @@ class W3CClient implements ClientInterface
      */
     public function setActiveTab(string $sessionIdentifier, string $tabIdentifier): PromiseInterface
     {
-        // TODO: Implement setActiveTab() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/window',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $requestContents = json_encode(['handle' => $tabIdentifier]);
+
+        $responsePromise = $this->httpClient->post($requestUri, $requestHeaders, $requestContents);
+
+        $switchConfirmationPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $this->onCommandConfirmation($response, 'Tab switch command is not confirmed.');
+
+                    return null;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to focus a tab.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $switchConfirmationPromise;
     }
 
     /**
@@ -222,9 +318,38 @@ class W3CClient implements ClientInterface
      */
     public function openUri(string $sessionIdentifier, string $uri): PromiseInterface
     {
-        // TODO: Implement openUri() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/url',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $requestContents = json_encode(['url' => $uri]);
+
+        $responsePromise = $this->httpClient->post($requestUri, $requestHeaders, $requestContents);
+
+        $navigationPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $this->onCommandConfirmation($response, 'URI navigation is not confirmed.');
+
+                    return null;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to open an URI.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $navigationPromise;
     }
 
     /**
@@ -232,9 +357,36 @@ class W3CClient implements ClientInterface
      */
     public function getCurrentUri(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement getUri() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/url',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $responsePromise = $this->httpClient->get($requestUri, $requestHeaders);
+
+        $currentUriPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $uriCurrent = $this->deserializeResponse($response);
+
+                    return $uriCurrent;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to get a current URI.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $currentUriPromise;
     }
 
     /**
@@ -242,9 +394,36 @@ class W3CClient implements ClientInterface
      */
     public function getSource(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement getSource() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/source',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $responsePromise = $this->httpClient->get($requestUri, $requestHeaders);
+
+        $sourceCodePromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $sourceCode = $this->deserializeResponse($response);
+
+                    return $sourceCode;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to get source code of the web resource.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $sourceCodePromise;
     }
 
     /**
@@ -252,19 +431,79 @@ class W3CClient implements ClientInterface
      */
     public function getElementIdentifier(string $sessionIdentifier, string $xpathQuery): PromiseInterface
     {
-        // TODO: Implement getElementIdentifier() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/element',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $requestContents = json_encode(['using' => 'xpath', 'value' => $xpathQuery]);
+
+        $responsePromise = $this->httpClient->post($requestUri, $requestHeaders, $requestContents);
+
+        $elementIdentifierPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $elementIdentifier = $this->extractElementIdentifier($response);
+
+                    return $elementIdentifier;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to get an element identifier.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $elementIdentifierPromise;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getActiveElement(string $sessionIdentifier): PromiseInterface
+    public function getActiveElementIdentifier(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement getActiveElement() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/element/active',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $responsePromise = $this->httpClient->get($requestUri, $requestHeaders);
+
+        $elementIdentifierPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $elementIdentifier = $this->extractElementIdentifier($response);
+
+                    return $elementIdentifier;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException(
+                        'Unable to get an identifier of the active element.',
+                        0,
+                        $rejectionReason
+                    );
+                }
+            )
+        ;
+
+        return $elementIdentifierPromise;
     }
 
     /**
@@ -272,9 +511,43 @@ class W3CClient implements ClientInterface
      */
     public function getElementVisibility(string $sessionIdentifier, array $elementIdentifier): PromiseInterface
     {
-        // TODO: Implement getElementVisibility() method.
+        // todo: safer checks (or hide internals behind a transfer object/contract)
+        $elementHandle = array_key_first($elementIdentifier);
+        if (!is_string($elementHandle)) {
+            throw new RuntimeException('Unexpected format for the element identifier.');
+        }
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/element/%s/displayed',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier,
+            $elementIdentifier[$elementHandle]
+        );
+
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $responsePromise = $this->httpClient->get($requestUri, $requestHeaders);
+
+        $visibilityStatusPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $visibilityStatus = $this->deserializeResponse($response);
+
+                    return $visibilityStatus;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to get visibility status for the element.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $visibilityStatusPromise;
     }
 
     /**
@@ -282,7 +555,7 @@ class W3CClient implements ClientInterface
      */
     public function clickElement(string $sessionIdentifier, array $elementIdentifier): PromiseInterface
     {
-        // TODO: Implement clickElement() method.
+        // todo: implementation
 
         return reject(new RuntimeException('Not implemented.'));
     }
@@ -295,9 +568,45 @@ class W3CClient implements ClientInterface
         array $elementIdentifier,
         string $keySequence
     ): PromiseInterface {
-        // TODO: Implement keypressElement() method.
+        // todo: safer checks (or hide internals behind a transfer object/contract)
+        $elementHandle = array_key_first($elementIdentifier);
+        if (!is_string($elementHandle)) {
+            throw new RuntimeException('Unexpected format for the element identifier.');
+        }
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/element/%s/value',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier,
+            $elementIdentifier[$elementHandle]
+        );
+
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $requestContents = json_encode(['text' => $keySequence]);
+
+        $responsePromise = $this->httpClient->post($requestUri, $requestHeaders, $requestContents);
+
+        $keypressConfirmationPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $this->onCommandConfirmation($response, 'Keypress command is not confirmed.');
+
+                    return null;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to apply keyboard keys to the element.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $keypressConfirmationPromise;
     }
 
     /**
@@ -310,9 +619,41 @@ class W3CClient implements ClientInterface
         int $moveDuration = 100,
         array $startingPoint = null
     ): PromiseInterface {
-        // TODO: Implement mouseMove() method.
+        if (is_array($startingPoint)) {
+            $actionOrigin = $startingPoint;
+        } else {
+            $actionOrigin = 'pointer';
+        }
 
-        return reject(new RuntimeException('Not implemented.'));
+        $mouseActions = [
+            [
+                'type'     => 'pointerMove',
+                'origin'   => $actionOrigin,
+                'x'        => $offsetX,
+                'y'        => $offsetY,
+                'duration' => $moveDuration,
+            ],
+        ];
+
+        $responsePromise = $this->requestMouseActions($sessionIdentifier, $mouseActions);
+
+        $moveConfirmationPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $this->onCommandConfirmation($response, 'Mouse move command is not confirmed.');
+
+                    return null;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to confirm mouse move action.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $moveConfirmationPromise;
     }
 
     /**
@@ -320,9 +661,36 @@ class W3CClient implements ClientInterface
      */
     public function mouseLeftClick(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement mouseLeftClick() method.
+        $mouseActions = [
+            [
+                'type'   => 'pointerDown',
+                'button' => 0,
+            ],
+            [
+                'type'   => 'pointerUp',
+                'button' => 0,
+            ],
+        ];
 
-        return reject(new RuntimeException('Not implemented.'));
+        $responsePromise = $this->requestMouseActions($sessionIdentifier, $mouseActions);
+
+        $clickConfirmationPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $this->onCommandConfirmation($response, 'Mouse click command is not confirmed.');
+
+                    return null;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to confirm mouse click action.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $clickConfirmationPromise;
     }
 
     /**
@@ -330,8 +698,148 @@ class W3CClient implements ClientInterface
      */
     public function getScreenshot(string $sessionIdentifier): PromiseInterface
     {
-        // TODO: Implement getScreenshot() method.
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/screenshot',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
 
-        return reject(new RuntimeException('Not implemented.'));
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $responsePromise = $this->httpClient->get($requestUri, $requestHeaders);
+
+        $imageContentsPromise = $responsePromise
+            ->then(
+                function (ResponseInterface $response) {
+                    $imageContentsEncoded = $this->deserializeResponse($response);
+                    $imageContents        = base64_decode($imageContentsEncoded);
+
+                    return $imageContents;
+                }
+            )
+            ->then(
+                null,
+                function (Throwable $rejectionReason) {
+                    throw new RuntimeException('Unable to get a screenshot.', 0, $rejectionReason);
+                }
+            )
+        ;
+
+        return $imageContentsPromise;
+    }
+
+    /**
+     * Initializes a request to execute a sequence of mouse-specific actions in the remote browser
+     *
+     * @param string $sessionIdentifier Session identifier for Selenium Grid server (hub)
+     * @param array  $mouseActions      A data structure that describes a sequence of actions to be performed by the
+     *                                  internal pointer with type "mouse"
+     *
+     * @return PromiseInterface<ResponseInterface>
+     */
+    private function requestMouseActions(string $sessionIdentifier, array $mouseActions): PromiseInterface
+    {
+        $requestUri = sprintf(
+            'http://%s:%d/wd/hub/session/%s/actions',
+            $this->_options['server']['host'],
+            $this->_options['server']['port'],
+            $sessionIdentifier
+        );
+
+        $requestHeaders = [
+            'Content-Type' => 'application/json; charset=UTF-8',
+        ];
+
+        $requestContents = json_encode(
+            [
+                'actions' => [
+                    [
+                        'type'       => 'pointer',
+                        'id'         => 'mouse',
+                        'parameters' => [
+                            'pointerType' => 'mouse',
+                        ],
+                        'actions'    => $mouseActions,
+                    ],
+                ],
+            ]
+        );
+
+        $responsePromise = $this->httpClient->post($requestUri, $requestHeaders, $requestContents);
+
+        return $responsePromise;
+    }
+
+    /**
+     * Returns an element identifier, which has to be extracted from the response message (a surgical approach)
+     *
+     * @param ResponseInterface $response PSR-7 response message from the Selenium hub with action results
+     *
+     * @return array
+     */
+    private function extractElementIdentifier(ResponseInterface $response): array
+    {
+        $responseBody = (string) $response->getBody();
+
+        preg_match(
+            '/(element(?:-[a-z\d]{4}){4}[a-z\d]{8})[":\s]+([a-z\d]{8}(?:-[a-z\d]{4}){4}[a-z\d]{8})/Ui',
+            $responseBody,
+            $matches
+        );
+
+        if (!isset($matches[1], $matches[2])) {
+            // todo: locate an error message or set it as "undefined error"
+            throw new RuntimeException('Unable to locate element identifier parts in the response.');
+        }
+
+        $elementIdentifier = [$matches[1] => $matches[2]];
+
+        return $elementIdentifier;
+    }
+
+    /**
+     * Ensures that a related action is properly executed (confirmed) by the remote server, triggers an error otherwise.
+     *
+     * It is used when no specific context is required to confirm successful command execution (some methods will use
+     * more advanced confirmation checks instead of this "default").
+     *
+     * @param ResponseInterface $response     PSR-7 response message from the Selenium hub with action results
+     * @param string            $errorMessage Will be used if an error is registered during confirmation check
+     *
+     * @return void
+     *
+     * @throws RuntimeException Whenever an error has been occurred during confirmation check for a remote action
+     */
+    private function onCommandConfirmation(ResponseInterface $response, string $errorMessage): void
+    {
+        $responseValueNode = $this->deserializeResponse($response);
+
+        // todo: locate an error message or set it as "undefined error"
+        if (null !== $responseValueNode) {
+            throw new RuntimeException($errorMessage);
+        }
+    }
+
+    /**
+     * Returns a "value" node contents, which will be extracted from the PSR-7 response message
+     *
+     * @param ResponseInterface $response PSR-7 response message from the Selenium hub with action results
+     *
+     * @return mixed
+     */
+    private function deserializeResponse(ResponseInterface $response)
+    {
+        $responseBody     = (string) $response->getBody();
+        $bodyDeserialized = json_decode($responseBody, true);
+
+        if (!array_key_exists('value', $bodyDeserialized)) {
+            // todo: locate an error message or set it as "undefined error"
+            throw new RuntimeException('Unable to locate "value" node (response deserialization).');
+        }
+
+        return $bodyDeserialized['value'];
     }
 }