+++ src/ClientInterface.php
@@ -174,7 +174,7 @@ interface ClientInterface
      *
      * @return PromiseInterface<array>
      */
-    public function getActiveElement(string $sessionIdentifier): PromiseInterface;
+    public function getActiveElementIdentifier(string $sessionIdentifier): PromiseInterface;
 
     /**
      * Returns a promise that resolves to a boolean, representing element visibility status on the currently active
@@ -193,8 +193,10 @@ interface ClientInterface
      * Returns a promise that will be resolved when the remote WebDriver service confirms element click operation
      * against active browser tab.
      *
-     * Note: if target element is a link, it may trigger an implicit tab(s) creation. In such case you need to take into
-     * consideration potentially new list of window handles for the current session (see {@link getTabIdentifiers()}).
+     * Note: if target element is a link, it may trigger an implicit tab(s) creation. In such case you need to take
+     * into consideration potentially new list of window handles for the current session (see {@link
+     * getTabIdentifiers()}). The remote webdriver service WILL NOT automatically perform tab switch operation if
+     * a command implicitly triggers new tab.
      *
      * @param string $sessionIdentifier Session identifier for Selenium Grid server (hub)
      * @param array  $elementIdentifier An internal WebDriver handle that refers to the element on the page