+++ src/WebDriverFactory.php
@@ -108,8 +108,8 @@ final class WebDriverFactory
                             'Maximum time to wait (in seconds) for command execution '
                             . '(do not correlate with HTTP timeouts)'
                         )
-                        ->allowedTypes('int')
-                        ->default(30)
+                        ->allowedTypes('float')
+                        ->default(30.0)
                     ;
                 }
             )
@@ -119,6 +119,10 @@ final class WebDriverFactory
 
         $socketConnector = new SocketConnector($loop, $optionsResolved['browser']);
         $httpClient      = new Browser($loop, $socketConnector);
+
+        // Selenium hub sends some valid responses with 5xx status codes, so we need to disable eager promise rejection
+        // to properly parse error messages and other details from the body.
+        $httpClient = $httpClient->withRejectErrorResponse(false);
 
         $hubClient          = new W3CClient($httpClient, ['server' => $optionsResolved['hub']]);
         $timeoutInterceptor = new TimeoutInterceptor($loop, $optionsResolved['command']['timeout']);